IAgentConnection: move to XI
